fix: Resolved another circular dependency
